+++ Changelog.md
@@ -1,5 +1,5 @@
 ## 0.2.2 (TBD)
-- Changes go here
+- Revert Fix `-` character must now be escaped in expression strings
 
 ## 0.2.1 (2021-01-02) [UtilityBeltInstaller-0.2.1.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/8552636461a1e2d1ec0c8c140cc7c9dd/UtilityBeltInstaller-0.2.1.exe)
 - Disabled character options profiles -- too buggy