+++ Changelog.md
@@ -7,6 +7,7 @@
 - Add missing vtank meta view expressions: `uigetcontrol`, `uisetlabel`, `uisetvisible`
 - Fix `InventoryManager.WatchLootProfile` not always watching for changes / double loading
 - Add settings to Nametags to adjust text sizes
+- UB expression patch now properly patches VTOptGet and VTOptSet meta actions
 
 ## 0.2.1 (2021-01-02) [UtilityBeltInstaller-0.2.1.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/8552636461a1e2d1ec0c8c140cc7c9dd/UtilityBeltInstaller-0.2.1.exe)
 - Disabled character options profiles -- too buggy