+++ Changelog.md
@@ -12,6 +12,7 @@
 - Add `getworldname[]` expression to return the name of the currently connected world / server
 - Add Alias tool for manipulating chat input. See https://utilitybelt.gitlab.io/docs/tools/aliases/
 - Add GameEvents tool for running actions when certain game events happen
+- Fix issue that was causing vtank funkiness (broken autovendor / keep # loot rules)
 
 ## 0.2.1 (2021-01-02) [UtilityBeltInstaller-0.2.1.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/8552636461a1e2d1ec0c8c140cc7c9dd/UtilityBeltInstaller-0.2.1.exe)
 - Disabled character options profiles -- too buggy