+++ Changelog.md
@@ -1,3 +1,6 @@
+## 0.2.3 (TBD)
+- Changes go here
+
 ## 0.2.2 (2021-01-16) [UtilityBeltInstaller-0.2.2.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/a597bed6c0330841d223ce2a762ca99d/UtilityBeltInstaller-0.2.2.exe)
 - Revert Fix `-` character must now be escaped in expression strings
 - Fix hotkeys not working